Added monitor changes and uploadmodifydelete js
These files are for adding, modifying, and deleting data from the database.  The uploadmodifydelete.js contains boiler plate variables that will be deleted and in their place will be accessors that access the actual form and use that data instead.  

The monitorchanges.js file is a work in progress and basically listens for changes to the database.